remove animation references
