+++ HackIllinois/Controllers/ViewControllers/HIEventViewControllers/HIEventListViewController.swift
@@ -35,11 +35,11 @@ extension HIEventListViewController {
 extension HIEventListViewController {
     // FIXME: remove after finishing layout debugging
     override func numberOfSections(in tableView: UITableView) -> Int {
-        return 3
+        return 1
     }
 
     override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return 4
+        return 5
     }
 
     override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
@@ -51,14 +51,6 @@ extension HIEventListViewController {
         }
         return cell
     }
-
-    func tableView(_ tableView: UITableView, viewForHeaderInSection section: Int) -> UIView? {
-        let header = tableView.dequeueReusableHeaderFooterView(withIdentifier: HIDateHeader.IDENTIFIER)
-        if let header = header as? HIDateHeader {
-            header.titleLabel.text = "\(section + 1):00 PM"
-        }
-        return header
-    }
 }
 
 // MARK: - UITableViewDelegate
@@ -66,7 +58,7 @@ extension HIEventListViewController {
     func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
         // FIXME: Remove
         return HIEventCell.heightForCell(with: indexPath.row)
-        guard let event = _fetchedResultsController?.object(at: indexPath) as? Event else {
+        guard let _ = _fetchedResultsController?.object(at: indexPath) as? Event else {
             return CGFloat.leastNonzeroMagnitude
         }
         return 0 // HIEventCell.heightForCell(displaying: event)
@@ -77,18 +69,10 @@ extension HIEventListViewController {
         eventDetailViewController.modalPresentationStyle = .overCurrentContext
         eventDetailViewController.modalTransitionStyle = .crossDissolve
         // eventDetailViewController.model = self._fetchedResultsController?.object(at: indexPath) as? Event
-//        eventDetailViewController.transitioningDelegate = self
+
         present(eventDetailViewController, animated: true, completion: nil)
         super.tableView(tableView, didSelectRowAt: indexPath)
     }
-}
-
-// MARK: - UIViewControllerTransitioningDelegate
-extension HIEventListViewController: UIViewControllerTransitioningDelegate {
-    func animationController(forPresented presented: UIViewController, presenting: UIViewController, source: UIViewController) -> UIViewControllerAnimatedTransitioning? {
-        return HIAnimator()
-    }
-
 }
 
 // MARK: - UIRefreshControl
